sui_swarm

Module memory

An in-memory, or rather in-process, backend for building and managing Sui Networks that all run inside the same process. Nodes are isolated from one another by each being run on their own separate thread within their own tokio runtime. This enables the ability to properly shut down a single node and ensure that all of its running tasks are also shut down, something that is extremely difficult or down right impossible to do if all the nodes are running on the same runtime.

Structs

Node

A handle to an in-memory Sui Node.

Swarm

A handle to an in-memory Sui Network.

SwarmBuilder

Enums

RuntimeType

The type of tokio runtime that should be used for a particular Node